feat: Separate hq-interop and http3 protocols into distinct implementations
- Add quicd-hq-interop crate for HTTP/0.9 over QUIC protocol
- Implement HqInteropApplication with plain text request parsing
- Update application registry to route hq-interop ALPNs correctly
- Clean up quicd-h3 to be pure HTTP/3 implementation
- Fix interop test handshake failure (✕(H) → ✓(H))
- Update configuration system and Docker build for new crate

+++ quicd-h3/src/frame.rs
@@ -311,7 +311,7 @@ impl FrameParser {
                 ParserState::ReadingLength { frame_type } => {
                     if let Some(length) = try_read_varint(&mut self.buffer)? {
                         eprintln!("FrameParser: Read frame length: {} for type 0x{:x}", length, frame_type);
-                        if length > u64::MAX as u64 {
+                        if length > u64::MAX {
                             return Err(Error::protocol(
                                 ErrorCode::FrameError,
                                 format!("frame length too large: {}", length),
